+++ src/theme/slideover.ts
@@ -18,30 +18,44 @@ const safeList = [
 
 export default {
   slots: {
-    overlay: 'fixed inset-0 bg-[#00204e]/52', //  bg-base-950/20 dark:bg-base-950/30 // [#00204e]/52
+    overlay: 'fixed inset-0 bg-[#00204e]/52',
     content: [
       'fixed',
-      'bg-base-50 dark:bg-base-950',
-      // 'divide-y divide-(--ui-border)',
+      'light --ui-context-content-light',
       'sm:shadow-lg',
       'flex flex-col focus:outline-none',
       'h-full'
     ].join(' '),
+    sidebarLayoutRoot: '',
+    sidebarLayoutPageBottomWrapper: '',
     header: [
-      // fix 'mt-4 px-5',
       'pt-[24px]',
       'flex items-center gap-1.5'
     ].join(' '),
-    wrapper: 'min-h-2xl',
+    wrapper: 'min-h-[30px]',
     body: [
       'size-full',
-      // fix 'mx-0 mt-2',
-      // fix 'flex-1 overflow-y-auto'
       'flex-1'
     ].join(' '),
-    footer: 'bg-white dark:bg-base-950 flex items-center justify-center gap-3 py-4 border-t border-t-1 border-t-base-900/10 dark:border-t-white/20 shadow-top-md p-2 pr-(--scrollbar-width)',
-    title: 'font-[family-name:var(--ui-font-family-primary)] text-(--ui-color-design-plain-content) font-semibold mb-0 text-(length:--ui-font-size-4xl) leading-5',
-    description: 'mt-2 mb-1 text-base-500 dark:text-base-400 text-sm',
+    footer: [
+      'bg-(--ui-color-bg-content-primary)',
+      'flex items-center justify-center gap-3',
+      'border-t border-t-1 border-t-(--ui-color-divider-less)',
+      'shadow-top-md',
+      'py-[9px] px-2 pr-(--scrollbar-width)'
+    ].join(' '),
+    title: [
+      'font-[family-name:var(--ui-font-family-primary)]',
+      'text-(--ui-color-design-plain-content)',
+      'font-semibold',
+      'mb-0',
+      'text-(length:--ui-font-size-4xl)/[calc(var(--ui-font-size-4xl)+2px)]'
+    ].join(' '),
+    description: [
+      'mt-1',
+      'text-(--ui-color-design-plain-na-content)',
+      'text-sm'
+    ].join(' '),
     close: 'absolute',
     safeList
   },
@@ -62,44 +76,31 @@ export default {
       right: { content: 'right-0 inset-y-0 w-[calc(100%-60px)] sm:w-[calc(100%-150px)]' },
       left: { content: 'left-0 inset-y-0 w-[calc(100%-60px)] sm:w-[calc(100%-150px)]' },
       top: {
-        // fix content: 'inset-x-0 top-0 max-h-full max-h-[calc(100%-18px)]'
-        content: [
-          'right-[5px] sm:right-[70px] bottom-0 sm:bottom-[18px] top-0',
-          'w-[calc(100%-60px-5px)] sm:w-[calc(100%-150px-70px)]',
-          'sm:max-h-[calc(100%-18px)]',
-          'sm:rounded-b-[18px]'
-        ].join(' ')
+        content: 'inset-x-0 top-0 max-h-full'
       },
       bottom: {
-        // fix content: 'inset-x-0 bottom-0 max-h-full'
         content: [
           'right-[5px] sm:right-[70px] top-0 sm:top-[18px] bottom-0',
           'w-[calc(100%-60px-5px)] sm:w-[calc(100%-150px-70px)]',
-          'sm:max-h-[calc(100%-18px)]',
-          'sm:rounded-t-[18px]'
-        ].join(' ')
+          'sm:max-h-[calc(100%-18px)]'
+        ].join(' '),
+        sidebarLayoutRoot: 'sm:rounded-t-[18px]'
       }
     },
     transition: {
       true: {
         overlay: 'motion-safe:data-[state=open]:animate-[fade-in_200ms_ease-out] motion-safe:data-[state=closed]:animate-[fade-out_200ms_ease-in]'
       }
-    },
-    // @todo remove this
-    scrollbarThin: {
-      true: {
-        // fix body: 'scrollbar-thin'
-      }
     }
   },
   compoundVariants: [
     // region btn.close ////
     {
-      side: ['right', 'top', 'bottom'],
+      side: ['right', 'bottom'],
       class: {
         close: [
           'pl-1.5 pr-[4px]',
-          'top-[17px] -translate-x-full left-0',
+          'top-[17px] -translate-x-full -left-px',
           'rounded-l-full'
         ].join(' ')
       }
@@ -110,7 +111,16 @@ export default {
         close: [
           'pr-1.5 pl-[4px]',
           'translate-x-full right-0',
-          'rounded-r-full'
+          'rounded-r-full',
+          '[&>div]:flex-row-reverse'
+        ].join(' ')
+      }
+    },
+    {
+      side: 'top',
+      class: {
+        close: [
+          'top-4 end-4'
         ].join(' ')
       }
     },